UI Improvements (#1)
* try adding the uploading message

* correctly apply state binding modifiers

* watch state changes

* fix $watch

* define dispatchFormEvent

* log more state changes

* use array instead of object for state

* track dispatching the form event

with console.logs of course

* fix check for files in progress

* remove console.logs

* better empty state

* more fixes for empty state

* Update README.md

* more style fixes

* show errors on uploads

* stop "uploading" spinner on errors

+++ resources/js/filament-uppy-component.js
@@ -10,6 +10,8 @@ window.fileUploaderComponent = function fileUploaderComponent({
     successEndpoint,
     errorEndpoint,
     deleteEndpoint,
+
+    uploadingMessage,
 }) {
     return {
         state,
@@ -23,8 +25,7 @@ window.fileUploaderComponent = function fileUploaderComponent({
         uppy: new Uppy({ autoProceed: true, allowMultipleUploads: true }),
 
         init() {
-            console.log('fileUploaderComponent init');
-            this.state = {};
+            this.state = [];
 
             window.addEventListener('beforeunload', (e) => {
                 if (!this.busy) return;
@@ -41,16 +42,25 @@ window.fileUploaderComponent = function fileUploaderComponent({
                         size: file.size,
                         progress: 0,
                     };
+
+                    this.dispatchFormEvent('form-processing-started', {
+                        message: uploadingMessage,
+                    })
                 })
                 .on('upload-progress', (file, progress) => this.filesInProgress[file.id].progress = ((progress.bytesUploaded / progress.bytesTotal) * 100).toFixed(0))
                 .on('upload-success', (file, response) => {
                     this.removeFileInProgress(file.id);
-                    this.state[file.id] = {
-                        id: file.id,
-                        name: file.name,
-                        size: file.size,
-                        url: response.uploadURL,
-                    };
+                    this.toggleFormProcessingState();
+
+                    // If state array does not contain a file with the same id, add it.
+                    if (!this.state.find((stateFile) => stateFile.id === file.id)) {
+                        this.state.push({
+                            id: file.id,
+                            name: file.name,
+                            size: file.size,
+                            url: response.uploadURL,
+                        });
+                    }
 
                     if (!!successEndpoint) {
                         const key = response.uploadURL.split('/').pop();
@@ -70,6 +80,7 @@ window.fileUploaderComponent = function fileUploaderComponent({
                 .on('upload-error', (file, error, response) => {
                     this.filesInProgress[file.id].error = true;
                     this.busy = true;
+                    this.toggleFormProcessingState();
 
                     if (!!errorEndpoint) {
                         fetch(errorEndpoint, {
@@ -109,20 +120,23 @@ window.fileUploaderComponent = function fileUploaderComponent({
         },
 
         removeCompletedFile(id) {
-            const key = this.state[id].url.split('/').pop();
-            const uuid = key.split('.')[0];
+            const fileIndex = this.state.findIndex((file) => file.id === id);
+            if (fileIndex !== -1) {
+                const key = this.state[fileIndex].url.split('/').pop();
+                const uuid = key.split('.')[0];
 
-            delete this.state[id];
+                this.state.splice(fileIndex, 1);
 
-            if (!!deleteEndpoint) {
-                fetch(deleteEndpoint, {
-                    method: 'DELETE',
-                    headers: {
-                        'Content-Type': 'application/json',
-                        'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]').content,
-                    },
-                    body: JSON.stringify({uuid}),
-                });
+                if (!!deleteEndpoint) {
+                    fetch(deleteEndpoint, {
+                        method: 'DELETE',
+                        headers: {
+                            'Content-Type': 'application/json',
+                            'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]').content,
+                        },
+                        body: JSON.stringify({ uuid }),
+                    });
+                }
             }
         },
 
@@ -139,6 +153,33 @@ window.fileUploaderComponent = function fileUploaderComponent({
 
         humanReadableFilesize(bytes) {
             return prettierBytes(bytes);
-        }
+        },
+
+        toggleFormProcessingState() {
+            let uploadsInProgress = false;
+            for (const file of this.uppy.getFiles()) {
+                if (file.progress.bytesUploaded < file.progress.bytesTotal) {
+                    uploadsInProgress = true;
+                }
+            }
+
+            if (uploadsInProgress) {
+                this.dispatchFormEvent('form-processing-started', {
+                    message: this.uploadingMessage,
+                });
+            } else {
+                this.dispatchFormEvent('form-processing-finished');
+            }
+        },
+
+        dispatchFormEvent: function (name, detail = {}) {
+            this.$el.closest('form')?.dispatchEvent(
+                new CustomEvent(name, {
+                    composed: true,
+                    cancelable: true,
+                    detail,
+                }),
+            )
+        },
     }
 }
